feat: 게시글 수정, 삭제 시 비밀번호 일치 확인

+++ src/main/java/com/sparta/crud/service/BoardService.java
@@ -9,9 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
@@ -23,7 +21,7 @@ public class BoardService {
     public BoardResponseDto createBoard(BoardRequestDto requestDto) {
         Board board = new Board(requestDto);
         boardRepository.save(board);
-        return new BoardResponseDto(board);
+        return new BoardResponseDto(board, true);
     }
 
     @Transactional(readOnly = true)
@@ -31,7 +29,7 @@ public class BoardService {
         List<Board> boardList = boardRepository.findAllByOrderByCreatedAtDesc();
         List<BoardResponseDto> boardResponseDtoList = new ArrayList<>();
         for (Board board : boardList) {
-            boardResponseDtoList.add(new BoardResponseDto(board));
+            boardResponseDtoList.add(new BoardResponseDto(board, true));
         }
         return boardResponseDtoList;
     }
@@ -41,7 +39,7 @@ public class BoardService {
         Board board = boardRepository.findById(id).orElseThrow(
                 () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
         );
-        return new BoardResponseDto(board);
+        return new BoardResponseDto(board, true);
     }
 
 //    @Transactional
@@ -55,24 +53,39 @@ public class BoardService {
         Board savedBoard = boardRepository.save(board);
 
         // 여기서 비밀번호 일치하는지 확인
+        if (requestDto.getPw().equals(board.getPw())) {
+            // 비밀번호 일치
 
+            // board 객체를 입력 받은 값으로 변경
+            savedBoard.update(requestDto);
+            // board를 저장
+            Board resultBoard = boardRepository.save(savedBoard);
+            // ResponseDto로 반환
+            return new BoardResponseDto(resultBoard, true);
+        } else {
+            // 비밀번호 불일치
 
-        // board 객체를 입력 받은 값으로 변경
-        savedBoard.update(requestDto);
+            return new BoardResponseDto(false);
+        }
 
-        // board를 저장
-        Board resultBoard = boardRepository.save(savedBoard);
-
-        // ResponseDto로 반환
-        return new BoardResponseDto(resultBoard);
     }
 
     @Transactional
-    public Map<String, Object> deleteBoard(Long id) {
-        Map<String, Object> response = new HashMap<>();
-        boardRepository.deleteById(id);
-        response.put("success", true);
-        return response;
-    }
+    public BoardResponseDto deleteBoard(Long id, BoardRequestDto requestDto) {
+        Board board = boardRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("아이디가 존재하지 않습니다.")
+        );
+        if (requestDto.getPw().equals(board.getPw())) {
+            // 비밀번호 일치
 
+            // 게시글 삭제
+            boardRepository.deleteById(id);
+            // ResponseDto로 반환
+            return new BoardResponseDto(true);
+        } else {
+            // 비밀번호 불일치
+
+            return new BoardResponseDto(false);
+        }
+    }
 }
